Tests/LibWeb/Text/input/css/css-math-empty-constructors.html

<!doctype html>
<script src="../include.js"></script>
<script>
    test(() => {
        for (const constructor of [
            CSSMathMax,
            CSSMathMin,
            CSSMathProduct,
            CSSMathSum,
        ]) {
            try {
                new constructor();
            } catch (e) {
                println(`${constructor.name}: ${e.name}`);
            }
        }
    });
</script>
