Tests/LibWeb/Text/expected/css/css-math-empty-constructors.txt

CSSMathMax: SyntaxError
CSSMathMin: SyntaxError
CSSMathProduct: SyntaxError
CSSMathSum: SyntaxError
